Restructure internal API classes/files
