+++ core/api/src/mill/api/internal/BaseModuleApi.scala
@@ -1,3 +1,5 @@
 package mill.api.internal
 
+import mill.api.internal.ModuleApi
+
 trait BaseModuleApi extends ModuleApi
