+++ core/api/src/mill/api/internal/EvaluatorApi.scala
@@ -1,9 +1,9 @@
 package mill.api.internal
 
 import mill.api.*
-
 import scala.collection.mutable
 import scala.util.DynamicVariable
+
 trait EvaluatorApi extends AutoCloseable {
   def evaluate(
       scriptArgs: Seq[String],
