+++ core/api/src/mill/api/internal/ModuleApi.scala
@@ -1,0 +1,9 @@
+package mill.api.internal
+
+import mill.api.Segments
+
+trait ModuleApi {
+  def moduleDirectChildren: Seq[ModuleApi]
+  private[mill] def moduleDirJava: java.nio.file.Path
+  def moduleSegments: Segments
+}
