+++ core/api/src/mill/api/internal/NamedTaskApi.scala
@@ -1,0 +1,5 @@
+package mill.api.internal
+
+trait NamedTaskApi[+T] extends TaskApi[T] {
+  def label: String
+}
