+++ core/api/src/mill/api/internal/PathRefApi.scala
@@ -1,0 +1,7 @@
+package mill.api.internal
+
+trait PathRefApi {
+  private[mill] def javaPath: java.nio.file.Path
+  def quick: Boolean
+  def sig: Int
+}
