+++ core/api/src/mill/api/internal/ScalaJSModuleApi.scala
@@ -1,0 +1,3 @@
+package mill.api.internal
+
+trait ScalaJSModuleApi extends JavaModuleApi
