+++ core/api/src/mill/api/internal/ScalaModuleApi.scala
@@ -1,0 +1,3 @@
+package mill.api.internal
+
+trait ScalaModuleApi extends JavaModuleApi
