+++ core/api/src/mill/api/internal/ScalaNativeModuleApi.scala
@@ -1,0 +1,3 @@
+package mill.api.internal
+
+trait ScalaNativeModuleApi extends JavaModuleApi
