+++ core/api/src/mill/api/internal/TaskApi.scala
@@ -1,0 +1,5 @@
+package mill.api.internal
+
+trait TaskApi[+T] {
+  def apply(): T
+}
