+++ core/api/src/mill/api/internal/UnresolvedPathApi.scala
@@ -1,0 +1,5 @@
+package mill.api.internal
+
+trait UnresolvedPathApi[P] {
+  def resolve(outPath: P): P
+}
