+++ core/api/src/mill/api/internal/bsp/BspBuildTarget.scala
@@ -1,4 +1,4 @@
-package mill.api.internal
+package mill.api.internal.bsp
 
 case class BspBuildTarget(
     displayName: Option[String],
