+++ core/api/src/mill/api/internal/bsp/BspModuleApi.scala
@@ -1,0 +1,30 @@
+package mill.api.internal.bsp
+
+import mill.api.internal.{ModuleApi, TaskApi}
+
+trait BspModuleApi extends ModuleApi {
+  private[mill] def bspBuildTargetData: TaskApi[Option[(String, AnyRef)]]
+  private[mill] def bspBuildTarget: BspBuildTarget
+  private[mill] def bspDisplayName: String
+}
+
+object BspModuleApi {
+
+  /** Used to define the [[BspBuildTarget.languageIds]] field. */
+  object LanguageId {
+    val Java = "java"
+    val Scala = "scala"
+    val Kotlin = "kotlin"
+  }
+
+  /** Used to define the [[BspBuildTarget.tags]] field. */
+  object Tag {
+    val Library = "library"
+    val Application = "application"
+    val Test = "test"
+    val IntegrationTest = "integration-test"
+    val Benchmark = "benchmark"
+    val NoIDE = "no-ide"
+    val Manual = "manual"
+  }
+}
