+++ core/api/src/mill/api/internal/bsp/BspServerResult.scala
@@ -1,4 +1,4 @@
-package mill.api.internal
+package mill.api.internal.bsp
 
 sealed trait BspServerResult
