+++ core/api/src/mill/api/internal/bsp/JvmBuildTarget.scala
@@ -1,4 +1,6 @@
-package mill.api.internal
+package mill.api.internal.bsp
+
+import mill.api.internal.bsp.BspUri
 
 case class JvmBuildTarget(
     javaHome: Option[BspUri],
