+++ core/api/src/mill/api/internal/idea/JavaFacet.scala
@@ -1,0 +1,3 @@
+package mill.api.internal.idea
+
+final case class JavaFacet(`type`: String, name: String, config: Element)
