+++ core/api/src/mill/api/internal/idea/ResolvedModule.scala
@@ -1,0 +1,23 @@
+package mill.api.internal.idea
+
+import mill.api.Segments
+import mill.api.internal.JavaModuleApi
+
+final case class ResolvedModule(
+    path: Segments,
+    classpath: Seq[Scoped[java.nio.file.Path]],
+    module: JavaModuleApi,
+    pluginClasspath: Seq[java.nio.file.Path],
+    scalaOptions: Seq[String],
+    scalaCompilerClasspath: Seq[java.nio.file.Path],
+    libraryClasspath: Seq[java.nio.file.Path],
+    facets: Seq[JavaFacet],
+    configFileContributions: Seq[IdeaConfigFile],
+    compilerOutput: java.nio.file.Path,
+    scalaVersion: Option[String],
+    resources: Seq[java.nio.file.Path],
+    generatedSources: Seq[java.nio.file.Path],
+    allSources: Seq[java.nio.file.Path]
+)
+
+
