+++ core/api/src/mill/api/internal/idea/Scoped.scala
@@ -1,0 +1,3 @@
+package mill.api.internal.idea
+
+final case class Scoped[T](value: T, scope: Option[String])
