+++ core/define/src/mill/define/BaseModule.scala
@@ -1,4 +1,5 @@
 package mill.define
+
 import mill.api.internal.BaseModuleApi
 
 /**
