+++ core/exec/src/mill/exec/GroupExecution.scala
@@ -12,7 +12,7 @@ import java.util.concurrent.ThreadPoolExecutor
 import scala.collection.mutable
 import scala.util.control.NonFatal
 import scala.util.hashing.MurmurHash3
-import mill.api.internal.{EvaluatorApi, BaseModuleApi, CompileProblemReporter, TestReporter}
+import mill.api.internal.{BaseModuleApi, CompileProblemReporter, EvaluatorApi, TestReporter}
 
 /**
  * Logic around evaluating a single group, which is a collection of [[Task]]s
