+++ integration/ide/gen-idea/resources/extended/build.mill
@@ -1,6 +1,6 @@
 import mill._
+import mill.api.internal.idea.{Element, IdeaConfigFile, JavaFacet}
 import mill.define.PathRef
-import mill.api.internal.{JavaFacet, IdeaConfigFile, Element}
 import mill.scalalib.TestModule
 
 trait HelloWorldModule extends scalalib.ScalaModule {
