+++ libs/kotlinlib/src/mill/kotlinlib/KotlinModule.scala
@@ -12,13 +12,14 @@ import mill.api.Result
 import mill.define.ModuleRef
 import mill.kotlinlib.worker.api.{KotlinWorker, KotlinWorkerTarget}
 import mill.scalalib.api.{CompilationResult, JvmWorkerApi}
-import mill.api.internal.{BspBuildTarget, BspModuleApi, CompileProblemReporter, internal}
+import mill.api.internal.{CompileProblemReporter, internal}
 import mill.scalalib.{JavaModule, JvmWorkerModule, Lib}
 import mill.util.Jvm
 import mill.*
 import java.io.File
 
 import mainargs.Flag
+import mill.api.internal.bsp.{BspBuildTarget, BspModuleApi}
 
 trait KotlinModule extends JavaModule { outer =>
