+++ runner/bsp/worker/src/mill/bsp/worker/BspClientType.scala
@@ -1,0 +1,17 @@
+package mill.bsp.worker
+
+/** Used to handle edge cases for specific BSP clients. */
+private[mill] enum BspClientType {
+
+  /** Intellij IDEA */
+  case IntellijBSP
+
+  /** Any other BSP client */
+  case Other(displayName: String)
+
+  def mergeResourcesIntoClasses: Boolean =
+    this match {
+      case IntellijBSP => true
+      case Other(_) => false
+    }
+}
