+++ runner/bsp/worker/src/mill/bsp/worker/Utils.scala
@@ -11,17 +11,11 @@ import ch.epfl.scala.bsp4j.{
   TaskId
 }
 import mill.api.ExecResult.{Skipped, Success}
-import mill.api.internal.{
-  TaskApi,
-  JavaModuleApi,
-  BspBuildTarget,
-  BspModuleApi,
-  ExecutionResultsApi,
-  CompileProblemReporter
-}
+import mill.api.internal.{CompileProblemReporter, ExecutionResultsApi, JavaModuleApi, TaskApi}
 
 import scala.jdk.CollectionConverters.*
 import scala.util.chaining.scalaUtilChainingOps
+import mill.api.internal.bsp.{BspBuildTarget, BspModuleApi}
 
 private object Utils {
